backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-PyMongo
Flask-Bcrypt
python-dotenv
langchain
cohere
black